+++ crates/holochain/src/core/workflow/validation_receipt_workflow/tests.rs
@@ -10,6 +10,7 @@ use holochain_types::prelude::*;
 use rusqlite::Transaction;
 
 #[tokio::test(flavor = "multi_thread")]
+#[ignore = "flaky"]
 async fn test_validation_receipt() {
     let _g = observability::test_run().ok();
     const NUM_CONDUCTORS: usize = 3;